adding new script for building search list by querying endpoint as opposed to bucket
